+++ src/renderer/renderer.js
@@ -1,12 +1,7 @@
-import {
-  // math
-  toRad,
-  // webgl
-  getGL, glErrnoToMsg,
-  // other
-  LoaderMerge
-} from '../utils.js';
-import {GameComponent} from '../game_component.js';
+import { toRad } from "../utils/math.js";
+import { getGL, glErrnoToMsg } from "../utils/gl_utils.js";
+import { LoaderMerge } from '../utils/loader.js';
+import { GameComponent } from '../game_component.js';
 
 import {Buffers} from './buffers.js';
 import {AtlasLoader} from './atlas_data.js';
